Increase Google TTS timeout to 15 seconds.
Default `HttpRequestBuilder` timeout is set to 5 seconds. When longer string
is passed to Google TTS or when Google servers are busy, the response is not
received during the 5 seconds.

This commit increases timeout to 15 seconds. This timeout is sufficient
based on my tests and I think it should not break anything.